+++ src/index.js
@@ -16,6 +16,7 @@ import _ComponentStyle from './models/ComponentStyle'
 import _styled from './constructors/styled'
 import _keyframes from './constructors/keyframes'
 import _injectGlobal from './constructors/injectGlobal'
+import _createGlobalStyle from './constructors/createGlobalStyle'
 import _constructWithOptions from './constructors/constructWithOptions'
 
 /* Import components */
@@ -76,6 +77,7 @@ const StyledComponent = _StyledComponent(ComponentStyle, constructWithOptions)
 /* Instantiate exported singletons */
 const keyframes = _keyframes(generateAlphabeticName, stringifyRules, css)
 const injectGlobal = _injectGlobal(stringifyRules, css)
+const createGlobalStyle = _createGlobalStyle(stringifyRules, css)
 const styled = _styled(StyledComponent, constructWithOptions)
 
 /* Export everything */
@@ -83,6 +85,7 @@ const styled = _styled(StyledComponent, constructWithOptions)
 export * from './secretInternals'
 export default styled
 export {
+  createGlobalStyle,
   css,
   keyframes,
   injectGlobal,